requirement.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.1
colorama==0.4.4
cycler==0.11.0
Flask==2.1.1
Flask-Cors==3.0.10
fonttools==4.31.2
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
tifffile==2022.3.25
torch==1.11.0
torchvision==0.12.0
tushare==1.2.84
typing_extensions==4.1.1
urllib3==1.26.9
websocket-client==0.57.0
Werkzeug==2.1.0
wincertstore==0.2
zipp==3.7.0
